+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fallback PNG codec via chunky_png
 - Rename `dither!` to `palette_reduce!`
 - Rename `#set_each_pixel_by_location` to `#set_each_pixel_by_location!` since it's mutable
+- ImageMagick codec can now read and write `gif` and `apng` including animations
 
 ## [0.2.0] - 2025-07-21
 - Ruby Sixel encoder now sets pixel aspect ratio metadata to display correctly in Windows Terminal